# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Layblar Platform API
  description: "# Introduction\nThis is the API Documentation for the Layblar-Platform.\
    \ \n"
  contact:
    name: Layblar Documentation
    url: https://layblar.github.io/docs
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
  x-logo:
    url: https://layblar.github.io/docs/img/layblarlogo.svg
servers:
- url: "http://{tenant}"
  variables:
    tenant:
      default: localhost:8080
...
paths:
...
  /household/{householdId}:
    get:
      tags:
      - Household
      summary: Get household information
      description: Get information about the given household
      operationId: getHouseholdInformation
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the data should be fetched
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Household with the given id
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
  /household/{householdId}/device:
    get:
      tags:
      - Household
      summary: List household devices
      description: List household devices
      operationId: listHouseholdDevices
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the data should be fetched
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: List of all devices in the household
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdDeviceDTO'
      security:
      - jwt: []
    put:
      tags:
      - Household
      summary: Update device information
      description: Update device information
      operationId: updateDeviceInformation
      parameters:
      - name: householdId
        in: path
        description: The id of the household from which the device should be removed
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The updated device
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                  deviceName:
                    type: string
                  deviceDescription:
                    type: string
                  manufacturer:
                    type: string
                  modelNumber:
                    type: string
                  powerDraw:
                    format: int32
                    type: integer
                  energyEfficiencyRating:
                    type: string
                  weight:
                    format: float
                    type: number
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceCategoryDTO'
      security:
      - jwt: []
    post:
      tags:
      - Household
      summary: Add device to household
      description: Add device to household
      operationId: addDeviceToHousehold
      parameters:
      - name: householdId
        in: path
        description: The id of the household to which the device should be added
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDeviceDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The household with the added device
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
  /household/{householdId}/device/{deviceId}:
    delete:
      tags:
      - Household
      summary: Remove device from household
      description: Remove device from household
      operationId: removeDeviceFromHousehold
      parameters:
      - name: deviceId
        in: path
        description: The id of the device that should be removed from the household
        required: true
        schema:
          type: string
      - name: householdId
        in: path
        description: The id of the household from which the device should be removed
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The household without the removed device
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                  deviceName:
                    type: string
                  deviceDescription:
                    type: string
                  manufacturer:
                    type: string
                  modelNumber:
                    type: string
                  powerDraw:
                    format: int32
                    type: integer
                  energyEfficiencyRating:
                    type: string
                  weight:
                    format: float
                    type: number
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceCategoryDTO'
      security:
      - jwt: []
  /household/{householdId}/join:
    get:
      tags:
      - Household
      summary: Get household join code
      description: Get the code for joining the given household
      operationId: getHouseholdJoinCode
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the join code should be fetched
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: JoinCode Information
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  joinCode:
                    type: string
                  validTill:
                    $ref: '#/components/schemas/LocalDateTime'
      security:
      - jwt: []
    post:
      tags:
      - Household
      summary: Join household
      description: Join given household
      operationId: joinHousehold
      parameters:
      - name: householdId
        in: path
        description: The id of the household that should be merged
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinHouseholdDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Your household
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
  /household/{householdId}/leave:
    post:
      tags:
      - Household
      summary: Leave household
      description: Leave given household
      operationId: leaveHousehold
      parameters:
      - name: householdId
        in: path
        description: The id of the household that should be split
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveHouseholdDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: Your household
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
  /household/{householdId}/smartMeterReader:
    post:
      tags:
      - Household
      summary: Register smart meter reader
      description: Register smart meter reader in given household
      operationId: registerSmartMeterReader
      parameters:
      - name: householdId
        in: path
        description: The id of the household to which the smart meter reader should
          be added
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartMeterDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The household with the added smart meter reader
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
    delete:
      tags:
      - Household
      summary: Remove smart meter reader
      description: Remove smart meter reader from given household
      operationId: removeSmartMeterReader
      parameters:
      - name: householdId
        in: path
        description: The id of the household from which the smart meter reader should
          be removed
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartMeterDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The household without the removed smart meter reader
          content:
            application/json:
              schema:
                type: object
                properties:
                  householdId:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdUserDTO'
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/HouseholdDeviceDTO'
                  smartMeters:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmartMeterDTO'
      security:
      - jwt: []
...
  /project/project/{projectId}/household/{householdId}:
    post:
      tags:
      - Project
      summary: Join a Project
      description: Join a Project
      operationId: joinProject
      parameters:
      - name: householdId
        in: path
        description: The ID of the household that wants to join the project
        required: true
        schema:
          type: string
      - name: projectId
        in: path
        description: The ID of the project to join
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinProjectDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The joined Project Information
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectName:
                    type: string
                  projectDescription:
                    type: string
                  projectDataUseDeclartion:
                    type: string
                  startDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  endDate:
                    $ref: '#/components/schemas/LocalDateTime'
                  created:
                    $ref: '#/components/schemas/LocalDateTime'
                  metaData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMetaDataDTO'
                  labels:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabelDTO'
      security:
      - jwt: []
  /smartmeter/household/{householdId}:
    get:
      tags:
      - Smart-Meter
      summary: Get Smart-Meter data
      description: Get a List of Smart-Meter data from the specified time frame
      operationId: getSmartMeterData
      parameters:
      - name: householdId
        in: path
        description: The id of the household for which the data should be fetched
        required: true
        schema:
          type: string
      - name: from
        in: query
        description: The begin of the time intervall for which the data should be
          fetched. Needs to be a Unix Timestamp as String
        required: true
        schema:
          type: string
      - name: to
        in: query
        description: The end of the time intervall for which the data should be fetched.
          Needs to be a Unix Timestamp as String
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "400":
          description: Missing Query Parameters
        "422":
          description: Wrong Date Format
        "200":
          description: Smart-Meter data for the specified time period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SmartMeterDataDTO'
      security:
      - jwt: []
components:
  schemas:
...
    AddDeviceDTO:
      description: Add Device to Household
      required:
      - deviceName
      - categories
      type: object
      properties:
        deviceId:
          description: If not provided a new deviceId will be created
          type: string
        deviceName:
          description: Name of the device
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          description: At least one category must be provided
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
...
    DeviceCategoryDTO:
      type: object
      properties:
        deviceCategoryId:
          type: string
        deviceCategoryName:
          type: string
        deviceCategoryDescription:
          type: string
    DeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
    HouseholdDeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
    HouseholdUserDTO:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    JoinHouseholdDTO:
      type: object
      properties:
        householdId:
          type: string
        userId:
          type: string
        joinCode:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    JoinProjectDTO:
      type: object
      properties:
        householdId:
          type: string
        metaData:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMetaDataDTO'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceDTO'
    LabelDTO:
      type: object
      properties:
        labelId:
          type: string
        labelName:
          type: string
        labelDescription:
          type: string
        labelMethod:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
...
    LeaveHouseholdDTO:
      type: object
      properties:
        householdId:
          type: string
        userId:
          type: string
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    ProjectMetaDataDTO:
      type: object
      properties:
        metaDataId:
          type: string
        metaDataName:
          type: string
        type:
          type: string
        isRequired:
          type: boolean
...
    SmartMeterDTO:
      type: object
      properties:
        smartMeterId:
          type: string
    SmartMeterDataDTO:
      description: Smart Meter Data
      type: object
      properties:
        time:
          description: Time at which Smart-Meter data was recored
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          example: 2023-05-06T11:15:40
        sensorId:
          description: Id of the Smart-Meter that recored the meassurement
          type: string
          example: 123MF34FBDKAS
        "1.7.0":
          format: float
          description: This is the current positive active power in watts measured
            from Phase A.
          type: number
        "1.8.0":
          format: float
          description: This is the total positive active energy in watt-hours consumed
            or generated from Phase A.
          type: number
        "2.7.0":
          format: float
          description: This is the current negative active power in watts measured
            from Phase A. This could indicate generated energy.
          type: number
        "2.8.0":
          format: float
          description: This is the total negative active energy in watt-hours consumed
            or generated from Phase A.
          type: number
        "3.8.0":
          format: float
          description: This is the total positive reactive energy in var-hours consumed
            or generated from Phase A.
          type: number
        "4.8.0":
          format: float
          description: This is the total negative reactive energy in var-hours consumed
            or generated from Phase A
          type: number
        "16.7.0":
          format: float
          description: This is the difference between positive and negative active
            instantaneous power in watts from Phase A.
          type: number
        "31.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L1.
          type: number
        "32.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L1.
          type: number
        "51.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L2.
          type: number
        "52.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L2.
          type: number
        "71.7.0":
          format: float
          description: This is the current in amperes flowing in Phase L3.
          type: number
        "72.7.0":
          format: float
          description: This is the instantaneous voltage in volts in Phase L3.
          type: number
...
    UpdateDeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    jwt:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    none:
      type: http
      description: No Authentication
      scheme: No-Auth